De-truncate object pointer state and simplify the tab-swap machinery
Sweep of parallel porting streams. Object classes that stashed
GameObject pointers or player handles in int-width state (SH_staff,
sharpclaw, the ECSH shrine and cup, thorntail, swapstone, WORLDplanet,
SB_Galleon, Cloudrun, ARWArwing, GF_LevelCon and friends) now declare
them at pointer width, and Sfx_ResolveObjectSfxId takes the
GameObject** it actually writes instead of laundering through an int
slot.

The per-tab fingerprint tracker in game_endian.c is replaced by
fhSwapLoadedTabFile, which swaps each tab file once at load time in
pi_dolphin.c; the resident-tab table walk and its heuristics go away
with it. The THP shim keys its RGB conversions off the attract movie
player's own texture set rather than shadow-tracking Y-plane pointers.

FOXHOLLOW_DEBUG_SHORTCUTS no longer alters game code at compile time,
and the debugging notes are updated for the current tooling.

+++ game/src/dlls/engine/23/23.c
@@ -440,10 +440,10 @@ int insertHighScore(u8 slot, u8 flag, u32 score, u8* initials)
 
             ((SaveScoreFile*)(saveData + off))->entries[rank].score = score;
             ((SaveScoreFile*)(saveData + off))->entries[rank].flag = flag;
-            ((SaveScoreFile*)((int)saveData + off))->entries[rank].initials[0] = initials[0];
-            ((SaveScoreFile*)((int)saveData + off))->entries[rank].initials[1] = initials[1];
-            ((SaveScoreFile*)((int)saveData + off))->entries[rank].initials[2] = initials[2];
-            ((SaveScoreFile*)((int)saveData + off))->entries[rank].initials[3] = initials[3];
+            ((SaveScoreFile*)(saveData + off))->entries[rank].initials[0] = initials[0];
+            ((SaveScoreFile*)(saveData + off))->entries[rank].initials[1] = initials[1];
+            ((SaveScoreFile*)(saveData + off))->entries[rank].initials[2] = initials[2];
+            ((SaveScoreFile*)(saveData + off))->entries[rank].initials[3] = initials[3];
             return rank;
         }
     }
