+++ game/src/dlls/objects/400_ECSH_Cup/ECSH_Cup.c
@@ -80,7 +80,7 @@ int ecshCup_getObjectTypeId(void) {
 }
 
 void ecshCup_free(GameObject* obj) {
-    (*gExpgfxInterface)->freeSource2((u32)obj);
+    (*gExpgfxInterface)->freeSource2((uintptr_t)obj);
 }
 
 void ecshCup_render(GameObject* obj, int renderArg2, int renderArg3, int renderArg4, int renderArg5, s8 visible) {
